refactor: rm unwanted
