--- src/index.ts
@@ -1,8 +1,0 @@
-async function main() {}
-
-main()
-  .then(() => process.exit(0))
-  .catch((e) => {
-    console.error(e);
-    process.exit(1);
-  });
